SACK receiver support
- not fully tested

+++ vnet/vnet/uri/uri_db.h
@@ -85,7 +85,7 @@ typedef CLIB_PACKED(struct
   u8 event_type;
   /* $$$$ for event logging */
   u16 event_id;
-  u16 enqueue_length;
+  u32 enqueue_length;
 }) fifo_event_t;
 
 typedef struct _stream_session_t
